Rename BasicEntityCheckBoxField to temp name
